split up cors lib and fix doc links
