+++ actix-cors/Cargo.toml
@@ -16,11 +16,11 @@ name = "actix_cors"
 path = "src/lib.rs"
 
 [dependencies]
-actix-web = { version = "3.0.0", default_features = false }
-actix-service = "1.0.6"
+actix-web = { version = "3.0.0", default-features = false }
 derive_more = "0.99.2"
 futures-util = { version = "0.3.4", default-features = false }
 
 [dev-dependencies]
+actix-service = "1.0.6"
 actix-rt = "1.1.1"
 regex = "1.3.9"
